🔥 Remove helpers.ETH2USD
