--- ad_source/tests/view_sets/__init__.py
@@ -1,3 +1,0 @@
-import json
-
-ETH2USD_DATA = json.dumps({"ETH": {"USD": 2206.09}, "BNB": {"USD": 496.74}, "MATIC": {"USD": 0.3474}})
