+++ ad_source/tests/view_sets/test_answer.py
@@ -3,8 +3,6 @@ from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-from ad_source.helpers import ETH2USD
-from ad_source.tests.view_sets import ETH2USD_DATA
 from ad_source.tests import mixins
 
 
@@ -32,7 +30,6 @@ class TestAnswerView(mixins.DataTestMixin, APITestCase):
 
     @responses.activate
     def test_create_answer_admin(self):
-        responses.add(responses.GET, ETH2USD.BASE_URL, body=ETH2USD_DATA, status=200)
         self.client.login(username='admin', password='admin')
         response = self.client.get(self.list_url)
         data = response.json()
@@ -108,7 +105,6 @@ class TestQuizAnswerView(mixins.DataTestMixin, APITestCase):
 
     @responses.activate
     def test_create_correct_admin(self):
-        responses.add(responses.GET, ETH2USD.BASE_URL, body=ETH2USD_DATA, status=200)
         self.client.login(username='admin', password='admin')
         response = self.client.get(self.list_url)
         data = response.json()
@@ -123,7 +119,6 @@ class TestQuizAnswerView(mixins.DataTestMixin, APITestCase):
     @responses.activate
     def test_create_wrong_admin(self):
         # Creating wrong answer is allowed. Task will get hidden
-        responses.add(responses.GET, ETH2USD.BASE_URL, body=ETH2USD_DATA, status=200)
         self.client.login(username='admin', password='admin')
         response = self.client.get(self.list_url)
         data = response.json()
