+++ ad_source/tests/view_sets/test_task_dashboard.py
@@ -6,8 +6,6 @@ from rest_framework import status
 from rest_framework.test import APITestCase
 
 from ad_source import models
-from ad_source.helpers import ETH2USD
-from ad_source.tests.view_sets import ETH2USD_DATA
 from ad_source.tests import mixins
 
 
@@ -34,7 +32,6 @@ class TestTaskDashboardView(mixins.DataTestMixin, APITestCase):
 
     @responses.activate
     def test_get_dashboard_list(self):
-        responses.add(responses.GET, ETH2USD.BASE_URL, body=ETH2USD_DATA, status=200)
         response = self.client.get(self.list_url)
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         data = response.json()
@@ -42,7 +39,6 @@ class TestTaskDashboardView(mixins.DataTestMixin, APITestCase):
 
     @responses.activate
     def test_get_dashboard_detail(self):
-        responses.add(responses.GET, ETH2USD.BASE_URL, body=ETH2USD_DATA, status=200)
         task = models.Task.objects.dashboard(self.admin).first()
         response = self.client.get(
             reverse('task_dashboard-detail', kwargs={'pk': task.id})
@@ -53,8 +49,6 @@ class TestTaskDashboardView(mixins.DataTestMixin, APITestCase):
 
     @responses.activate
     def test_update_dashboard_detail(self):
-        responses.add(responses.GET, ETH2USD.BASE_URL, body=ETH2USD_DATA, status=200)
-
         responses.add(responses.GET, 'https://example.com',
                       body=EXAMPLE_COM, status=200)
         data = {
@@ -71,7 +65,6 @@ class TestTaskDashboardView(mixins.DataTestMixin, APITestCase):
 
     @responses.activate
     def test_withdraw_dashboard_detail(self):
-        responses.add(responses.GET, ETH2USD.BASE_URL, body=ETH2USD_DATA, status=200)
         with patch('ad_source.tasks.update_task_is_active_balance.delay') as mock_update_task:
             task = models.Task.objects.dashboard(self.admin).last()
             response = self.client.post(
